created new webpack config for dev and production;
