+++ README.md
@@ -9,6 +9,13 @@ Run the following commands on the terminal after cloning the repository and movi
 2. npm run seed-database // adds data to the database
 3. npm start // starts server
 
+4. You must run either:
+  npm run react-dev for development mode
+  or
+  npm run webpack for production mode
+
+// this creates both the app.bundle.js and the vendor.bundle.js
+
 Then:
 Open the browser at localhost:4000
